SAK-26150 Support DisplayAdvisorUDP in JLDAPDirectoryProvider
Unless there's an attribute mapping for a displayId of displayName then the JLDAPDirectoryProvider just returns null which causes the UserDirectoryService to use the standard values. When a mapping attribute is supplied the value of that attribute is stored as a property on the user object and then used when the provider is asked for a displayId/displayName.

git-svn-id: https://source.sakaiproject.org/svn/providers/trunk@314082 66ffb92e-73f9-0310-93c1-f5514f145a0a